+++ src/main/java/com/diagnomind/web_server/domain/hospital/service/HospitalService.java
@@ -18,7 +18,7 @@ public class HospitalService {
 
     private final HospitalRepository hospitalRepository;
 
-    public Optional<User> addUser(Integer gid, User user) {
+    public Optional<User> addUser(Long gid, User user) {
         return hospitalRepository
                 .findById(gid)
                 .map(hospital -> {
@@ -27,7 +27,7 @@ public class HospitalService {
                 });
     }
 
-    public Optional<User> getUser(Integer gid, Integer uid) {
+    public Optional<User> getUser(Long gid, Long uid) {
         return hospitalRepository
                 .findById(gid)
                 .flatMap(hospital -> hospital
@@ -39,14 +39,14 @@ public class HospitalService {
                         .findFirst());
     }
 
-    public List<User> getAllUsers(Integer gid) {
+    public List<User> getAllUsers(Long gid) {
         return hospitalRepository
                 .findById(gid)
                 .map(Hospital::getUsers)
                 .orElse(List.of());
     }
 
-    public Optional<User> modifyUser(Integer gid, User modifiedUser) {
+    public Optional<User> modifyUser(Long gid, User modifiedUser) {
         return hospitalRepository
                 .findById(gid)
                 .flatMap(hospital -> hospital
@@ -59,7 +59,7 @@ public class HospitalService {
                         .map(foundUser -> foundUser.update(modifiedUser)));
     }
 
-    public boolean deleteUser(Integer gid, Integer uid) {
+    public boolean deleteUser(Long gid, Long uid) {
         return hospitalRepository
                 .findById(gid)
                 .map(hospital -> hospital
@@ -74,7 +74,7 @@ public class HospitalService {
         return hospitalRepository.save(hospital);
     }
 
-    public Optional<Hospital> getHospital(Integer gid) {
+    public Optional<Hospital> getHospital(Long gid) {
         return hospitalRepository.findById(gid);
     }
 
@@ -90,7 +90,7 @@ public class HospitalService {
                 .map(hospital -> hospital.update(modifiedHospital));
     }
 
-    public boolean deleteHospital(Integer gid) {
+    public boolean deleteHospital(Long gid) {
         hospitalRepository.deleteById(gid);
         return !hospitalRepository.existsById(gid);
     }
