+++ src/main/java/com/diagnomind/web_server/domain/training_data/repository/TrainingDataRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.diagnomind.web_server.domain.training_data.model.TrainingData;
 
-public interface TrainingDataRepository extends CrudRepository<TrainingData, Integer> {
+public interface TrainingDataRepository extends CrudRepository<TrainingData, Long> {
     
 }
